# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AuthKit API
  version: 0.1.0-rc.1
  description: Static integration contract. Runtime documentation endpoints are intentionally disabled.
servers:
  - url: https://auth.example.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/social/{providerKey}/start:
    post:
      security: []
      parameters: [{$ref: '#/components/parameters/ProviderKey'}]
      requestBody:
        content: {application/json: {schema: {$ref: '#/components/schemas/SocialLoginStartRequest'}}}
      responses: {'200': {description: 'OIDC authorization URL with server-bound state, nonce and PKCE'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/social/{providerKey}/callback:
    get:
      security: []
      parameters:
        - {$ref: '#/components/parameters/ProviderKey'}
        - {name: state, in: query, required: true, schema: {type: string, minLength: 32, maxLength: 256}}
        - {name: code, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 4096}}
      responses: {'200': {description: Federated authentication or explicit linking completed}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
...
  /api/v1/users/me/social-identities:
    get: *authGet
  /api/v1/users/me/social-identities/{providerKey}/link/start:
    post:
      parameters: [{$ref: '#/components/parameters/ProviderKey'}]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/StepUpRequest'}}}}
      responses: {'200': {description: OIDC authorization URL for an explicit link ceremony}, '401': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/social-identities/{identityId}:
    delete:
      parameters: [{$ref: '#/components/parameters/IdentityId'}]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/StepUpRequest'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {description: Last usable authenticator cannot be removed}}
  /api/v1/oauth2/authorize:
    description: Deprecated JSON compatibility endpoint. New clients use the conventional browser endpoint below.
    post:
      requestBody: {$ref: '#/components/requestBodies/OAuthAuthorize'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /oauth2/authorize:
    get:
      security: []
      description: Starts an Authorization Code flow and redirects to the operator login/consent UI with an opaque signed transaction handle.
      parameters:
        - {name: response_type, in: query, required: true, schema: {type: string, enum: [code]}}
        - {name: client_id, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
        - {name: redirect_uri, in: query, required: true, schema: {type: string, format: uri, maxLength: 512}}
        - {name: scope, in: query, required: true, schema: {type: string, minLength: 1}}
        - {name: state, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 255}}
        - {name: code_challenge, in: query, required: true, schema: {type: string, pattern: '^[A-Za-z0-9_-]{43,128}$'}}
        - {name: code_challenge_method, in: query, required: true, schema: {type: string, enum: [S256]}}
        - {name: nonce, in: query, required: false, schema: {type: string, maxLength: 255}, description: Required when scope contains openid}
      responses:
        '302': {description: 'Redirect to the operator UI, or a validated client redirect carrying a standard OAuth error', headers: {Location: {schema: {type: string, format: uri}}}}
        '400': {$ref: '#/components/responses/OAuthError'}
  /api/v1/oauth2/authorize/transactions/{transaction}:
    parameters:
      - {name: transaction, in: path, required: true, schema: {type: string, minLength: 32, maxLength: 512}}
    get:
      responses:
        '200': {description: Login and consent transaction details, content: {application/json: {schema: {$ref: '#/components/schemas/OAuthAuthorizationTransactionApiResponse'}}}}
        '400': {$ref: '#/components/responses/OAuthError'}
    post:
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/OAuthAuthorizationDecisionRequest'}}}
      responses:
        '302': {description: Exact registered client redirect with code/state or access_denied/state, headers: {Location: {schema: {type: string, format: uri}}}}
        '400': {$ref: '#/components/responses/OAuthError'}
  /oauth2/token:
    post:
      security: [{oauthClientBasic: []}]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: {$ref: '#/components/schemas/OAuthTokenRequest'}
      responses:
        '200': {description: Standard OAuth token response, content: {application/json: {schema: {$ref: '#/components/schemas/OAuthTokenResponse'}}}}
        '400': {$ref: '#/components/responses/OAuthError'}
        '401': {$ref: '#/components/responses/OAuthError'}
        '503': {description: Dependency unavailable}
  /oauth2/revoke:
    post:
      security: [{oauthClientBasic: []}]
      requestBody: {$ref: '#/components/requestBodies/OAuthRevocation'}
      responses: {'200': {description: Token revoked or already inactive}, '400': {$ref: '#/components/responses/OAuthError'}, '401': {$ref: '#/components/responses/OAuthError'}}
  /oauth2/introspect:
    post:
      security: [{oauthClientBasic: []}]
      requestBody: {$ref: '#/components/requestBodies/OAuthIntrospection'}
      responses: {'200': {description: Standard OAuth introspection response, content: {application/json: {schema: {$ref: '#/components/schemas/OAuthIntrospectionResponse'}}}}, '400': {$ref: '#/components/responses/OAuthError'}, '401': {$ref: '#/components/responses/OAuthError'}}
  /oauth2/userinfo:
    get:
      security: [{oauthAccessBearer: []}]
      description: Requires an AuthKit OAuth access token with `token_use=oauth_access`; ID tokens and first-party access tokens are rejected.
      responses: {'200': {description: OIDC userinfo claims}, '400': {description: Invalid token class or malformed bearer token}, '401': {description: Unauthorized}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: First-party AuthKit access token with `token_use=first_party_access`, configured API audience, non-revoked key, and active server-side session. OAuth access tokens and ID tokens are rejected.
    oauthAccessBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth access token with `token_use=oauth_access`, OAuth client audience, `client_id`, and server-side scope checks. ID tokens and first-party access tokens are rejected.
    refreshCookie:
      type: apiKey
      in: cookie
      name: Refresh-Token
      description: HttpOnly refresh cookie scoped to AuthKit refresh/logout routes.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: CSRF header that must echo the readable CSRF cookie on refresh and logout.
    oauthClientBasic:
      type: http
      scheme: basic
      description: OAuth client authentication for token, introspection, and revocation endpoints. Public clients may use approved body credentials where supported.
    workerToken:
      type: apiKey
      in: header
      name: X-Worker-Token
      description: Rotatable internal credential accepted only from explicitly trusted network peers.
  parameters:
...
    ProviderKey: {name: providerKey, in: path, required: true, schema: {type: string, pattern: '^[a-z0-9][a-z0-9_-]{1,63}$'}}
...
    IdentityId: {name: identityId, in: path, required: true, schema: {type: string, format: uuid}}
  requestBodies:
...
    OAuthAuthorize: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OAuthAuthorizeRequest'}}}}
    OAuthIntrospection: {required: true, content: {application/x-www-form-urlencoded: {schema: {$ref: '#/components/schemas/OAuthIntrospectionRequest'}}}}
    OAuthRevocation: {required: true, content: {application/x-www-form-urlencoded: {schema: {$ref: '#/components/schemas/OAuthRevocationRequest'}}}}
...
  headers:
    XRequestId:
      required: true
      schema: {type: string, format: uuid}
      description: Request correlation identifier; a valid inbound X-Request-ID is preserved, otherwise AuthKit generates one.
  responses:
    Success:
      description: Successful response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
    Error:
      description: Opaque error response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
    OAuthError:
      description: Standard OAuth protocol error; never wrapped in the AuthKit envelope
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/OAuthError'}}}
...
  schemas:
...
    OAuthAuthorizeRequest:
      type: object
      additionalProperties: false
      required: [responseType, clientId, redirectUri, scope, codeChallenge, codeChallengeMethod]
      properties:
        responseType: {type: string, enum: [code]}
        clientId: {type: string, minLength: 1, maxLength: 128}
        redirectUri: {type: string, format: uri, maxLength: 512}
        scope: {type: string, minLength: 1, maxLength: 512}
        state: {type: string, maxLength: 255, nullable: true}
        codeChallenge: {type: string, pattern: '^[A-Za-z0-9_-]{43,128}$'}
        codeChallengeMethod: {type: string, enum: [S256]}
        nonce: {type: string, maxLength: 255, nullable: true}
        consentAccepted: {type: boolean, nullable: true}
    OAuthIntrospectionRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: {type: string, minLength: 1, maxLength: 8192, writeOnly: true}
        token_type_hint: {type: string, enum: [access_token, refresh_token], nullable: true}
        client_id: {type: string, maxLength: 128, nullable: true}
        client_secret: {type: string, maxLength: 2048, nullable: true, writeOnly: true}
    OAuthRevocationRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: {type: string, minLength: 1, maxLength: 8192, writeOnly: true}
        token_type_hint: {type: string, enum: [access_token, refresh_token], nullable: true}
        client_id: {type: string, maxLength: 128, nullable: true}
        client_secret: {type: string, maxLength: 2048, nullable: true, writeOnly: true}
    OAuthIntrospectionResponse:
      type: object
      additionalProperties: false
      required: [active]
      properties:
        active: {type: boolean}
        sub: {type: string, format: uuid}
        client_id: {type: string}
        scope: {type: string}
        token_type: {type: string, enum: [Bearer]}
        exp: {type: integer, format: int64}
        iat: {type: integer, format: int64}
        iss: {type: string, format: uri}
        aud: {type: array, items: {type: string}}
        token_use: {type: string, enum: [oauth_access]}
...
    OAuthError:
      type: object
      additionalProperties: false
      required: [error, error_description]
      properties:
        error: {type: string, enum: [invalid_request, invalid_client, invalid_grant, invalid_token, unsupported_grant_type, unsupported_response_type, invalid_scope, access_denied]}
        error_description: {type: string}
    OAuthTokenRequest:
      type: object
      additionalProperties: false
      required: [grant_type]
      properties:
        grant_type: {type: string, enum: [authorization_code, refresh_token]}
        code: {type: string}
        redirect_uri: {type: string, format: uri}
        code_verifier: {type: string, pattern: '^[A-Za-z0-9._~-]{43,128}$'}
        refresh_token: {type: string, description: Opaque rotating bearer credential}
        client_id: {type: string}
        client_secret: {type: string, writeOnly: true}
    OAuthTokenResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in, scope]
      properties:
        access_token: {type: string, description: JWT with token_use=oauth_access}
        id_token: {type: string, nullable: true, description: JWT with token_use=id_token}
        refresh_token: {type: string, nullable: true, description: 'Opaque, rotating and family-bound'}
        token_type: {type: string, enum: [Bearer]}
        expires_in: {type: integer, format: int64, minimum: 1}
        scope: {type: string}
    OAuthAuthorizationDecisionRequest:
      type: object
      additionalProperties: false
      required: [approved]
      properties: {approved: {type: boolean}}
    OAuthAuthorizationTransactionApiResponse:
      type: object
      required: [data, timestamp, requestId]
      properties:
        data:
          type: object
          required: [clientId, clientDisplayName, scopes, consentRequired, expiresInSeconds]
          properties:
            clientId: {type: string}
            clientDisplayName: {type: string}
            scopes: {type: array, items: {type: string}}
            consentRequired: {type: boolean}
            expiresInSeconds: {type: integer, format: int64, minimum: 0}
        errors: {type: array, items: {type: string}, nullable: true}
        timestamp: {type: string, format: date-time}
        requestId: {type: string}
    ApiResponse:
      type: object
      additionalProperties: false
      required: [timestamp, requestId]
      properties:
        data: {}
        errors: {type: array, items: {type: string}, nullable: true}
        timestamp: {type: string, format: date-time}
        code: {type: string, pattern: '^[a-z0-9_]+$', nullable: true, description: Stable machine-readable error code}
        requestId: {type: string, minLength: 1, maxLength: 128}
...
    SocialLoginStartRequest:
      type: object
      additionalProperties: false
      properties:
        termsAccepted: {type: boolean, default: false}
        privacyPolicyAccepted: {type: boolean, default: false}
    StepUpRequest:
      type: object
      additionalProperties: false
      properties:
        currentPassword: {type: string, maxLength: 128, nullable: true}
        mfaCode: {type: string, minLength: 6, maxLength: 32, nullable: true}
...